+++ design-toolkit/components/src/components/combobox/index.tsx
@@ -10,6 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
+'use client';
 import 'client-only';
 import {
   Button as AriaButton,
@@ -30,7 +31,7 @@ import {
 } from 'react-aria-components';
 
 import { cn } from '@/lib/utils';
-import ChevronDown from '@accelint/icons/chevron-down';
+import { ChevronDown } from '@accelint/icons';
 import { type VariantProps, cva } from 'cva';
 import type { ReactNode } from 'react';
 import { Icon } from '../icon';
@@ -149,7 +150,7 @@ export function ComboBox<T extends IOptionsItem>({
             <Label
               className='empty:hidden'
               isDisabled={isDisabled}
-              isOptional={!props.isRequired}
+              isRequired={props.isRequired}
             >
               {label}
             </Label>