+++ design-toolkit/components/src/components/text-area/index.tsx
@@ -10,6 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
+'use client';
 import 'client-only';
 import {
   Text as AriaText,
@@ -174,7 +175,7 @@ export function TextArea({
         <Label
           className='empty:hidden'
           isDisabled={isDisabled}
-          isOptional={!props.isRequired}
+          isRequired={props.isRequired}
         >
           {label}
         </Label>